Update speakers carousel section
