+++ index.html
@@ -19,15 +19,15 @@
   <script src='https://cdn.jsdelivr.net/npm/@fullcalendar/google-calendar@6.1.15/index.global.min.js'></script>
   <script src='https://cdn.jsdelivr.net/npm/@fullcalendar/timegrid@6.1.15/index.global.min.js'></script>
 
-  <script type="text/javascript" src="//cdn.jsdelivr.net/npm/slick-carousel@1.8.1/slick/slick.min.js"></script>
 
   <!-- Carousel -->
-  <link rel="stylesheet" type="text/css" href="slick/slick.css" />
-  <link rel="stylesheet" type="text/css" href="slick/slick-theme.css" />
-  <link rel="stylesheet" type="text/css" href="//cdn.jsdelivr.net/npm/slick-carousel@1.8.1/slick/slick.css" />
+  <script defer type="text/javascript" src="//cdn.jsdelivr.net/npm/slick-carousel@1.8.1/slick/slick.min.js"></script>
+  <link defer rel="stylesheet" type="text/css" href="slick/slick.css" />
+  <link defer rel="stylesheet" type="text/css" href="slick/slick-theme.css" />
+  <link defer rel="stylesheet" type="text/css" href="//cdn.jsdelivr.net/npm/slick-carousel@1.8.1/slick/slick.css" />
   <script src="https://code.jquery.com/jquery-3.7.1.slim.min.js" integrity="sha256-kmHvs0B+OpCW5GVHUNjv9rOmY0IvSIRcf7zGUDTDQM8=" crossorigin="anonymous"></script>
-  
-  <script src="script/speakers.js" defer></script>
+
+  <script src="script/speakersCarousel.js" type="module" defer></script>
   <script src="script/matrix.js" defer></script>
   <script src="script/calendar.js" defer></script>
 </head>
@@ -55,9 +55,9 @@
           <li class="nav-item">
             <a class="nav-link" href="#program">Programação</a>
           </li>
-          <!-- <li class="nav-item">
-            <a class="nav-link" href="#palest">Palestrantes</a>
-          </li> -->
+          <li class="nav-item">
+            <a class="nav-link" href="#speakers">Palestrantes</a>
+          </li>
           <!-- <li class="nav-item">
             <a class="nav-link" href="#hackaton">Hackaton</a>
           </li>
@@ -159,7 +159,7 @@
     <section id="speakers" class="container mb-5 pb-3">
       <h2>Palestrantes</h2>
       <div id="carrossel" class="carrossel"></div>
-      <div class="modal fade" id="speakerModal" tabindex="-1" aria-labelledby="speakerModalLabel" aria-hidden="true">
+      <div class="modal modal-lg fade" id="speakerModal" tabindex="-1" aria-labelledby="speakerModalLabel" aria-hidden="true">
         <div class="modal-dialog">
           <div class="modal-content">
             <div class="modal-header">
@@ -174,15 +174,15 @@
               </div>
               <div id="modal-speaker-abstract">
                 <div class="mb-2 mt-3">
-                  <strong>Resumo</strong>
+                  <strong>Sobre mim</strong>
                 </div>
-                <div id="speaker-abstract" class="event-section"></div>
+                <div id="speaker-abstract"></div>
               </div>
               <div id="modal-speaker-lectures">
                 <div class="mb-2 mt-3">
                   <strong>Eventos</strong>
                 </div>
-                <div id="speaker-events" class="event-section"></div>
+                <ul id="speaker-events"></ul>
               </div>
             </div>
             <div id="modal-footer" class="modal-footer">
