+++ style/style.css
@@ -11906,15 +11906,6 @@ div.speaker-contact-card {
   height: 3rem;
 }
 
-footer {
-  margin-top: 2rem;
-  background-color: #2e2e5e;
-  color: white;
-  padding: 1rem 0;
-  text-align: center;
-  height: 5rem;
-}
-
 /* FullCalendar */
 .fc-col-header-cell-cushion {
   color: rgb(255, 255, 255);
@@ -11973,44 +11964,14 @@ div.event-section {
   border: 1px solid white;
 }
 
-/* Carrosel */
-.slick-slide > div {
-  margin-left: 5px;
-  margin-right: 5px;
-}
-
-div.carrossel-item {
-  padding: 10px;
-}
-
-.main-bg-color {
-  background-color: #2e2e5e;
-}
-.main-bg-color:focus {
-  background-color: #2e2e5e;
-}
-
-.main-border-color {
-  border-color: white;
-}
-
-.carrossel {
-  font-family: "Roboto Mono", monospace;
-  margin-left: 5%;
-  margin-right: 5%;
-  margin-top: 20px;
-}
-
+/* #region Carousel --------------------------------------------------------- */
 .carrossel-item {
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-  max-width: 300px;
-  margin: auto;
+  margin: 1rem;
+  padding: 2rem 1rem;
   text-align: center;
-  align-items: center;
-  color: var(--bs-nav-tabs-link-active-color);
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
 }
-
-.carrossel-item img {
+.carrossel-item > img {
   display: block;
   margin: 0 auto;
   margin-top: 20px;
@@ -12018,6 +11979,21 @@ div.carrossel-item {
   height: 100px;
   border-radius: 50%;
   object-fit: cover;
+}
+
+#speakerModal #modal-speaker-abstract,
+#speakerModal #modal-speaker-lectures {
+  padding: 1rem;
+  border-bottom: 1px solid rgba(255, 255, 255, 0.3);
+}
+#speakerModal #modal-speaker-lectures {
+  padding-bottom: 0;
+  border-bottom: none;
+}
+#speakerModal #modal-speaker-photo {
+  display: block;
+  width: 60%;
+  height: auto;
 }
 
 .modal-footer .contact-list {
@@ -12042,6 +12018,16 @@ div.carrossel-item {
 }
 .modal-footer .contact-list i {
   font-size: 30px;
+}
+
+/* #endregion Carousel ------------------------------------------------------ */
+footer {
+  margin-top: 2rem;
+  background-color: #2e2e5e;
+  color: white;
+  padding: 1rem 0;
+  text-align: center;
+  height: 5rem;
 }
 
 /*# sourceMappingURL=style.css.map */
