+++ style/style.scss
@@ -67,15 +67,6 @@ div.speaker-contact-card {
   height: 3rem;
 }
 
-footer {
-  margin-top: 2rem;
-  background-color: $primary;
-  color: white;
-  padding: 1rem 0;
-  text-align: center;
-  height: 5rem;
-}
-
 /* FullCalendar */
 
 .fc-col-header-cell-cushion {
@@ -129,55 +120,42 @@ div.event-section {
   border: 1px solid white;
 }
 
-/* Carrosel */
+/* #region Carousel --------------------------------------------------------- */
 
-.slick-slide > div {
-  margin-left: 5px;
-  margin-right: 5px;
-}
+.carrossel-item {
+  margin: 1rem;
+  padding: 2rem 1rem;
+  text-align: center;
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
 
-div.carrossel-item {
-  padding: 10px;
-}
-
-.main-bg-color {
-  background-color: $primary;
-
-  &:focus {
-    background-color: $primary;
+  > img {
+    display: block;
+    margin: 0 auto;
+    margin-top: 20px;
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    object-fit: cover;
   }
 }
 
-.main-border-color {
-  border-color: white;
-}
+#speakerModal {
+  #modal-speaker-abstract,
+  #modal-speaker-lectures {
+    padding: 1rem;
+    // padding-top: 0.5rem;
+    border-bottom: 1px solid rgba($color: #fff, $alpha: 0.3);
+  }
+  #modal-speaker-lectures {
+    padding-bottom: 0;
+    border-bottom: none;
+  }
 
-.carrossel {
-  font-family: "Roboto Mono", monospace;
-  margin-left: 5%;
-  margin-right: 5%;
-  margin-top: 20px;
-}
-
-.carrossel-item {
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-  max-width: 300px;
-  // height: 250px;
-  margin: auto;
-  text-align: center;
-  align-items: center;
-  // background-color: #343a40;
-  color: var(--bs-nav-tabs-link-active-color);
-}
-
-.carrossel-item img {
-  display: block;
-  margin: 0 auto;
-  margin-top: 20px;
-  width: 100px;
-  height: 100px;
-  border-radius: 50%;
-  object-fit: cover;
+  #modal-speaker-photo {
+    display: block;
+    width: 60%;
+    height: auto;
+  }
 }
 
 .modal-footer {
@@ -208,4 +186,16 @@ div.carrossel-item {
       font-size: 30px;
     }
   }
+}
+
+/* #endregion Carousel ------------------------------------------------------ */
+
+
+footer {
+  margin-top: 2rem;
+  background-color: $primary;
+  color: white;
+  padding: 1rem 0;
+  text-align: center;
+  height: 5rem;
 }
